Add SearchProfileCard
